applications/email_sender.py
import smtplib
import os
from email.message import EmailMessage
from typing import List, Optional
from pathlib import Path
from applications.logger import get_logger
from dotenv import load_dotenv

load_dotenv()

class EmailSender:
    """
    Handles sending job applications via SMTP email with attachments.
    """
    def __init__(self):
        self.smtp_server = os.getenv("SMTP_SERVER")
        self.smtp_port = int(os.getenv("SMTP_PORT", 587))
        self.smtp_user = os.getenv("SMTP_USER")
        self.smtp_password = os.getenv("SMTP_PASSWORD")
        self.logger = get_logger()

    def send_email(
        self,
        to_email: str,
        subject: str,
        body: str,
        attachments: Optional[List[Path]] = None,
        cc: Optional[List[str]] = None,
        bcc: Optional[List[str]] = None
    ) -> bool:
        """
        Send an email with optional attachments.
        Returns True if sent successfully, False otherwise.
        """
        msg = EmailMessage()
        msg["From"] = self.smtp_user
        msg["To"] = to_email
        msg["Subject"] = subject
        if cc:
            msg["Cc"] = ", ".join(cc)
        if bcc:
            msg["Bcc"] = ", ".join(bcc)
        msg.set_content(body)

        # Attach files
        if attachments:
            for file_path in attachments:
                try:
                    with open(file_path, "rb") as f:
                        file_data = f.read()
                        file_name = file_path.name
                    msg.add_attachment(file_data, maintype="application", subtype="octet-stream", filename=file_name)
                except Exception as e:
                    self.logger.error(f"Failed to attach {file_path}: {e}")

        try:
            with smtplib.SMTP(self.smtp_server, self.smtp_port) as server:
                server.starttls()
                server.login(self.smtp_user, self.smtp_password)
                server.send_message(msg)
            self.logger.info(f"Email sent to {to_email} with subject '{subject}'")
            return True
        except Exception as e:
            self.logger.error(f"Failed to send email to {to_email}: {e}")
            return False
